telegram/telegram-light-jobs/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.31.0
feedparser>=6.0.11
newspaper3k>=0.2.8
lxml_html_clean>=0.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
google-cloud-storage>=2.16.0
pymongo>=4.6.0
python-dotenv>=1.0.0
